news_api_app/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    SECRET_KEY: str = os.getenv("SECRET_KEY")
    ALGORITHM: str = os.getenv("ALGORITHM", "HS256")
    ACCESS_TOKEN_EXPIRE_MINUTES: int = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))
    DATABASE_URL: str = os.getenv("DATABASE_URL")
    NEWSAPI_KEY = "28433e4f84904eb2a27047d066f2d2c1"
    NEWSAPI_BASE_URL = "https://newsapi.org/v2"

settings = Settings()
