news_api_app/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
python-dotenv
pydantic
